requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
minio==7.2.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.25
Pillow==10.2.0
